Extract systemvm.iso using bsdtar if available.

+++ scripts/vm/systemvm/injectkeys.sh
@@ -6,9 +6,9 @@
 # to you under the Apache License, Version 2.0 (the
 # "License"); you may not use this file except in compliance
 # with the License.  You may obtain a copy of the License at
-# 
+#
 #   http://www.apache.org/licenses/LICENSE-2.0
-# 
+#
 # Unless required by applicable law or agreed to in writing,
 # software distributed under the License is distributed on an
 # "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -33,32 +33,56 @@ clean_up() {
   $SUDO umount $MOUNTPATH
 }
 
+clean_up_bsdtar() {
+  rm -rf --preserve-root $MOUNTPATH
+}
+
+backup_iso() {
+  $SUDO cp -b ${systemvmpath} ${systemvmpath}.bak
+}
+
 inject_into_iso() {
   local isofile=${systemvmpath}
   local newpubkey=$2
-  local backup=${isofile}.bak
   local tmpiso=${TMP}/$1
   mkdir -p $MOUNTPATH
   [ ! -f $isofile ] && echo "$(basename $0): Could not find systemvm iso patch file $isofile" && return 1
-  $SUDO mount -o loop $isofile $MOUNTPATH 
-  [ $? -ne 0 ] && echo "$(basename $0): Failed to mount original iso $isofile" && clean_up && return 1
-  diff -q $MOUNTPATH/authorized_keys $newpubkey &> /dev/null && clean_up && return 0
-  $SUDO cp -b $isofile $backup
-  [ $? -ne 0 ] && echo "$(basename $0): Failed to backup original iso $isofile" && clean_up && return 1
-  rm -rf $TMPDIR
-  mkdir -p $TMPDIR
-  [ ! -d $TMPDIR  ] && echo "$(basename $0): Could not find/create temporary dir $TMPDIR" && clean_up && return 1
-  $SUDO cp -fr $MOUNTPATH/* $TMPDIR/
-  [ $? -ne 0 ] && echo "$(basename $0): Failed to copy from original iso $isofile" && clean_up && return 1
-  $SUDO cp $newpubkey $TMPDIR/authorized_keys
-  [ $? -ne 0 ] && echo "$(basename $0): Failed to copy key $newpubkey from original iso to new iso " && clean_up && return 1
-  mkisofs -quiet -r -o $tmpiso $TMPDIR
-  [ $? -ne 0 ] && echo "$(basename $0): Failed to create new iso $tmpiso from $TMPDIR" && clean_up && return 1
-  $SUDO umount $MOUNTPATH
-  [ $? -ne 0 ] && echo "$(basename $0): Failed to unmount old iso from $MOUNTPATH" && return 1
-  $SUDO cp -f $tmpiso $isofile
-  [ $? -ne 0 ] && echo "$(basename $0): Failed to overwrite old iso $isofile with $tmpiso" && return 1
-  rm -rf $TMPDIR
+  if [ -x "$(command -v bsdtar)" ]; then
+    bsdtar -C $MOUNTPATH -xf $isofile
+    [ $? -ne 0 ] && echo "$(basename $0): Failed to extract original iso $isofile" && clean_up_bsdtar && return 1
+    diff -q $MOUNTPATH/authorized_keys $newpubkey &> /dev/null && clean_up_bsdtar && return 0
+    backup_iso
+    [ $? -ne 0 ] && echo "$(basename $0): Failed to backup original iso $isofile" && clean_up_bsdtar && return 1
+    $SUDO cp $newpubkey $MOUNTPATH/authorized_keys
+    [ $? -ne 0 ] && echo "$(basename $0): Failed to copy key $newpubkey from original iso to new iso " && clean_up_bsdtar && return 1
+    mkisofs -quiet -r -o $tmpiso $MOUNTPATH
+    [ $? -ne 0 ] && echo "$(basename $0): Failed to create new iso $tmpiso from $MOUNTPATH" && clean_up_bsdtar && return 1
+    $SUDO cp -f $tmpiso $isofile
+    [ $? -ne 0 ] && echo "$(basename $0): Failed to overwrite old iso $isofile with $tmpiso" && return 1
+    clean_up_bsdtar
+  else
+    $SUDO mount -o loop $isofile $MOUNTPATH
+    [ $? -ne 0 ] && echo "$(basename $0): Failed to mount original iso $isofile" && clean_up && return 1
+    diff -q $MOUNTPATH/authorized_keys $newpubkey &> /dev/null && clean_up && return 0
+    backup_iso
+    [ $? -ne 0 ] && echo "$(basename $0): Failed to backup original iso $isofile" && clean_up && return 1
+    #
+    rm -rf $TMPDIR
+    mkdir -p $TMPDIR
+    [ ! -d $TMPDIR  ] && echo "$(basename $0): Could not find/create temporary dir $TMPDIR" && clean_up && return 1
+    $SUDO cp -fr $MOUNTPATH/* $TMPDIR/
+    [ $? -ne 0 ] && echo "$(basename $0): Failed to copy from original iso $isofile" && clean_up && return 1
+    #
+    $SUDO cp $newpubkey $TMPDIR/authorized_keys
+    [ $? -ne 0 ] && echo "$(basename $0): Failed to copy key $newpubkey from original iso to new iso " && clean_up && return 1
+    mkisofs -quiet -r -o $tmpiso $TMPDIR
+    [ $? -ne 0 ] && echo "$(basename $0): Failed to create new iso $tmpiso from $TMPDIR" && clean_up && return 1
+    $SUDO umount $MOUNTPATH
+    [ $? -ne 0 ] && echo "$(basename $0): Failed to unmount old iso from $MOUNTPATH" && return 1
+    $SUDO cp -f $tmpiso $isofile
+    [ $? -ne 0 ] && echo "$(basename $0): Failed to overwrite old iso $isofile with $tmpiso" && return 1
+    rm -rf $TMPDIR
+  fi
 }
 
 copy_priv_key() {
@@ -74,7 +98,7 @@ then
    SUDO="sudo -n "
 fi
 
-$SUDO mkdir -p $MOUNTPATH
+mkdir -p $MOUNTPATH
 
 [ $# -ne 3 ] && echo "Usage: $(basename $0)  <new public key file> <new private key file> <systemvm iso path>" && exit 3
 newpubkey=$1
@@ -85,17 +109,8 @@ systemvmpath=$3
 
 command -v mkisofs > /dev/null   || (echo "$(basename $0): mkisofs not found, please install or ensure PATH is accurate" ; exit 4)
 
-# if running into Docker as unprivileges, skip ssh verification as iso cannot be mounted due to missing loop device.
-if [ -f /.dockerenv ]; then
-  if [ -e /dev/loop0 ]; then
-    # it's a docker instance with privileges.
-    inject_into_iso systemvm.iso $newpubkey
-    [ $? -ne 0 ] && exit 5
-    copy_priv_key $newprivkey
-  else
-    # this mean it's a docker instance, ssh key cannot be verify.
-    echo "We run inside Docker, skipping ssh key insertion in systemvm.iso"
-  fi
+if [ ! -e /dev/loop0 ] && [ ! -x "$(command -v bsdtar)" ]; then
+  echo "Loop device is missing and bsdtar is unavailable. Skipping ssh key insertion in systemvm.iso"
 else
   inject_into_iso systemvm.iso $newpubkey
   [ $? -ne 0 ] && exit 5
